refactor: proper naming for dropdown action creator
