+++ src/actions/index.ts
@@ -51,7 +51,7 @@ export { default as initThoughts } from './initThoughts'
 export { default as invalidState } from './invalidState'
 export { default as join } from './join'
 export { default as jump } from './jump'
-export { default as manageDropdowns } from './manageDropdowns'
+export { default as toggleDropDowns } from './toggleDropdowns'
 export { default as mergeThoughts } from './mergeThoughts'
 export { default as moveThought } from './moveThought'
 export { default as moveThoughtDown } from './moveThoughtDown'
