+++ src/components/Content.tsx
@@ -5,7 +5,7 @@ import Dispatch from '../@types/Dispatch'
 import SimplePath from '../@types/SimplePath'
 import { Thunk } from '../@types/Thunk'
 import { closeModalActionCreator as closeModal } from '../actions/closeModal'
-import { manageDropdownsActionCreator as closeAllDropdowns } from '../actions/manageDropdowns'
+import { toggleDropDownsActionCreator as closeAllDropdowns } from '../actions/toggleDropdowns'
 import { isTouch } from '../browser'
 import { ABSOLUTE_PATH, HOME_PATH, TUTORIAL2_STEP_SUCCESS } from '../constants'
 import { childrenFilterPredicate, filterAllChildren } from '../selectors/getChildren'
