+++ src/components/Editable.tsx
@@ -13,9 +13,9 @@ import { errorActionCreator as error } from '../actions/error'
 import importData from '../actions/importData'
 import { importSpeechToTextActionCreator as importSpeechToText } from '../actions/importSpeechToText'
 import { setInvalidStateActionCreator as setInvalidState } from '../actions/invalidState'
-import { manageDropdownsActionCreator as closeAllDropdowns } from '../actions/manageDropdowns'
 import { newThoughtActionCreator as newThought } from '../actions/newThought'
 import { setCursorActionCreator as setCursor } from '../actions/setCursor'
+import { toggleDropDownsActionCreator as closeAllDropdowns } from '../actions/toggleDropdowns'
 import { tutorialNextActionCreator as tutorialNext } from '../actions/tutorialNext'
 import { isMac, isTouch } from '../browser'
 import { commandEmitter } from '../commands'
